// src/types/supabase-generated.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instanciate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
// ...
      matches: {
        Row: {
          created_at: string | null
          date: string
          id: string
          location: string
          match_number: number | null
          player1_id: string
          player2_id: string
          round: number | null
          score: Json | null
          status: string | null
          summary: string | null
          tournament_id: string | null
          updated_at: string | null
          winner_id: string | null
        }
        Insert: {
          created_at?: string | null
          date: string
          id?: string
          location: string
          match_number?: number | null
          player1_id: string
          player2_id: string
          round?: number | null
          score?: Json | null
          status?: string | null
          summary?: string | null
          tournament_id?: string | null
          updated_at?: string | null
          winner_id?: string | null
        }
        Update: {
          created_at?: string | null
          date?: string
          id?: string
          location?: string
          match_number?: number | null
          player1_id?: string
          player2_id?: string
          round?: number | null
          score?: Json | null
          status?: string | null
          summary?: string | null
          tournament_id?: string | null
          updated_at?: string | null
          winner_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "fk_tournament"
            columns: ["tournament_id"]
            isOneToOne: false
            referencedRelation: "tournaments"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "matches_player1_id_fkey"
            columns: ["player1_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["user_id"]
          },
          {
            foreignKeyName: "matches_player2_id_fkey"
            columns: ["player2_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["user_id"]
          },
          {
            foreignKeyName: "matches_winner_id_fkey"
            columns: ["winner_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["user_id"]
          },
        ]
      }
// ...
      tournaments: {
        Row: {
          brackets_generated: boolean | null
          created_at: string | null
          description: string
          end_date: string
          entry_fee: number | null
          format: string | null
          id: string
          location: string
          max_participants: number
          name: string
          organizer_id: string
          prize_pool: number | null
          start_date: string
          status: string | null
          updated_at: string | null
        }
        Insert: {
          brackets_generated?: boolean | null
          created_at?: string | null
          description: string
          end_date: string
          entry_fee?: number | null
          format?: string | null
          id?: string
          location: string
          max_participants: number
          name: string
          organizer_id: string
          prize_pool?: number | null
          start_date: string
          status?: string | null
          updated_at?: string | null
        }
        Update: {
          brackets_generated?: boolean | null
          created_at?: string | null
          description?: string
          end_date?: string
          entry_fee?: number | null
          format?: string | null
          id?: string
          location?: string
          max_participants?: number
          name?: string
          organizer_id?: string
          prize_pool?: number | null
          start_date?: string
          status?: string | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "tournaments_organizer_id_fkey"
            columns: ["organizer_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["user_id"]
          },
        ]
      }
    }
// ...
  }
}